Feature EPL-1731: Using outlet instead of conditionals in home
