+++ packages/ComponentLibrary/src/components/PrimaryTab/types.ts
@@ -5,7 +5,6 @@ export interface TabItem {
   fill?: string;
   hoverFill?: string;
   showInTab: 'icon' | 'label' | 'both';
-  href: string;
 }
 
 export interface PrimaryTabsProps {
