+++ packages/MainUI/src/main.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import Home from './screens/Home';
-import Table from './screens/Table';
+import Table from './screens/Table/TableView';
 import Login from './screens/Login';
 import Form from './screens/Form';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
